+++ src/shared/api/index.ts
@@ -1,3 +1,2 @@
-export { useGetPostsForUser } from "./useGetPostsForUser";
 export { useGetTimelinePosts } from "./useGetTimelinePosts";
 export { useSearchPosts } from "./useSearchPosts";
